refactor: rename map/style.ts to mapbox-style.ts as suggested in PR review
